backend/requirements.txt
requests
pytest
sqlalchemy
pydantic
uvicorn
fastapi
python-dotenv
psycopg
